requirements.txt
# ========================================
# Core Dependencies
# ========================================
numpy==2.2.6
pillow==11.3.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0
setuptools==80.9.0
charset-normalizer==3.4.2
attrs==25.3.0
click==8.2.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
referencing==0.36.2
rpds-py==0.26.0

# ========================================
# ML & AI
# ========================================
torch
transformers==4.53.2
tokenizers==0.21.2
sentencepiece==0.2.0
safetensors==0.5.3
huggingface-hub==0.34.2
tiktoken==0.11.0
regex==2024.11.6
filelock==3.18.0
packaging==25.0
tqdm==4.67.1

# CUDA dependencies (for GPU support)
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
triton==3.3.1
sympy==1.14.0
mpmath==1.3.0
networkx==3.4.2
fsspec==2025.3.0

# ========================================
# OpenAI API
# ========================================
openai==1.99.5
pydub==0.25.1
anyio==4.9.0
distro==1.9.0
httpx==0.28.1
sniffio==1.3.1
h11==0.16.0
httpcore==1.0.9
certifi==2025.7.14
idna==3.10
typing_extensions==4.14.1
annotated-types==0.7.0
jiter==0.10.0

# ========================================
# FastAPI Backend
# ========================================
fastapi>=0.115.0
uvicorn[standard]>=0.31.1
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# FastAPI dependencies
starlette
anyio
httptools==0.6.4
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
bcrypt==4.3.0

# ========================================
# Neo4j Database
# ========================================
neo4j>=5.16.0

# ========================================
# Document Processing
# ========================================
google-cloud-documentai>=2.20.0
google-cloud-storage>=2.10.0
pymupdf>=1.23.0
opendataloader-pdf>=1.1.0

# Google Cloud dependencies
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
cachetools==5.5.2
requests-oauthlib==2.0.0
urllib3==2.5.0
oauthlib==3.3.1
